requirements.txt
pycryptodome==3.12.0
pyAesCrypt==6.0.0
logger-1.0.6 @ git+https://github.com/Sithi5/logger@1.0.6